Check if file exists before removing in clear.ps1

+++ clear.ps1
@@ -1,10 +1,16 @@
 Push-Location
 
 Set-Location -Path ".\nuget\win-x64"
-Remove-Item "hf_tokenizers.dll" -Force
-Remove-Item "Tokenizers.DotNet.runtime.win.*.nupkg" -Force
+if (Test-Path -Path "Tokenizers.DotNet.runtime.win.*.nupkg") {
+    Remove-Item "Tokenizers.DotNet.runtime.win.*.nupkg" -Force
+}
+if (Test-Path -Path "hf_tokenizers.dll") {
+    Remove-Item "hf_tokenizers.dll" -Force
+}
 
 Set-Location -Path ".."
-Remove-Item "Tokenizers.DotNet.*" -Force
+if (Test-Path -Path "Tokenizers.DotNet.*") {
+    Remove-Item "Tokenizers.DotNet.*" -Force
+}
 
 Pop-Location
